Move ValidatorTests to its own package
