+++ src/main/java/seedu/address/model/item/ItemQuantity.java
@@ -2,7 +2,7 @@ package seedu.address.model.item;
 
 import static java.util.Objects.requireNonNull;
 
-import seedu.address.model.item.itemvalidator.ItemQuantityValidator;
+import seedu.address.model.item.validators.ItemQuantityValidator;
 
 
 /**
